comm_dissector: KML - switched positioning to absolute.
I tried using relative positioning, as that would make the path
always visible and easier to use. But this cannot be done for
a multipart model, due to KML format restricions.

Ground altitude is now an export parameter; without it typed correctly,
the path will be either high in the air or under ground.