comm_dissector: KML - updated initial timestamps and stick refersh
Initial values of timestamps are not treated more suspicious, as they
sometimes consists of timestamps at time of production. These are now
replaced.

Stick refersh conditions have been moved to a function. It looks like
sometimes the sticks leave an unintended trace on screen; this makes
it easier to diagnose such issues.