comm_dissector: KML - rotated 3D model.
It turns out Google Earth can't properly handle rotations where roll is
near 90/-90 degrees. If it is 4 degrees around this value, it starts
rotating invalid axis, and at exact value the model gest rotated 180 in
yaw.

Why you do this, Google? This is math level at 1st year of high school.
And you keep such mistake in a software which is 13 years old?
Seroiusly, teach your interns to google, Google.

Anyway the model was rotated 90 degrees, so this issue was visible
almost all the time. Now I rotated the model so that usual rotation in
Google Earth will be 0 - this should make the issue less visible.