dji_imah_fwsig: Type field is 4 chars.
The field was previously unused, so we didn't knew for sure
what its format was. But in WM240 it is used - stores short
text name of the module.

The change will cause old .ini files to set type to one-char string
of '0' while signing; but no matter - it was unused anyway.
And this can be easily corrected by just generating new .ini
(unsigning the firmware file before re-signing).