dji_imah_fwsig: If cannot decrypt, extract with padding
When the tool was asked to force-extract encrypted data, it was cutting
the padding. That was stupid, and made the data impossible to fully
decrypt.

This patch makes it possible to extract encrypted data properly.
I doubt anyone would really want to extract encrypted data.
But at least now the SIG file can be re-created properly from these.