tests: Added Mavic Air, increased tolerance for padding
The signing tool still do not get padding right in all platforms.
Maybe DJI sometimes pads the firmwares with garbage?
The scheme seem to change between plaforms.
Not that this really matters - it's just padding.
But it does generate more binary differences between original
and re-signed files, so requires wider tolerance when testing.