comm_sbs_bqctrl: Change structure to allow re-setting globals
Global variables from main script are modified by chip drivers.
Without the special support to reset these, it was not possible to
run the main function of the script several times, which is
required for `pylabview` testing.

Now the script sha been rebuilt so that there is a function which resets
global variables.