tests: comm_sbs_bqctrl testing in pytest
It would be hard to introduce CI to the project while tests
are the simple bash scripts. Today, `pytest` is used to implement
tests in most projects - let's use that as well.

Starting with tests for `comm_sbs_bqctrl.py`. The tests are roughly
matching the shell script tests, though pytest provides a bit more
utility, so new ones might be better.