dji_imah_fwsig: Better support of modules without encryption
The modules for which encryption key fourcc was set to \0\0\0\0
were unsigninig and re-signing correctly, but:
1. were showing several warnings
2. were storing the scramble_key supposedly in encrypted form

So while this worked, the strange way of storing scramble_key
suggested that the output binary is encryprted and not usable.
That was not the case, because there was no encryption at all.

Since we use the scramble_key storage form as indicator of a
skewed output binary in our tests, this was causing all the
tests on unsigned binaries to skip.

This patch makes the scramble_key to be stored normally within
INI file, while also removing most of the unneccessary warnings.

There are no known IMaH v1 modules with encryption key set to
\0\0\0\0. This means this patch, in practice, only affacts
IMaH v2 files.